requirements.txt
nltk
boolean.py
